requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
black==22.3.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
coloredlogs==15.0.1
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==1.1.1
fairscale==0.4.12
filelock==3.8.0
flatbuffers==2.0.7
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.10.1
humanfriendly==10.0
hydra-core==1.2.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iopath==0.1.9
ipykernel==6.16.2
ipython==8.5.0
ipywidgets==8.0.2
jedi==0.18.1
jupyter-client==7.4.4
jupyter-core==4.11.2
jupyterlab-widgets==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.0
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.2
munch==2.5.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
-e git+https://github.com/Leonid-Erlygin/seminars.git@17f1ca2b20fc9b201a54244b8ef2d1dd57d37c28#egg=nn_interpretability&subdirectory=l5_visualization/vis_tools/nn_interpretability
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
platformdirs==2.5.2
portalocker==2.6.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pyDeprecate==0.3.1
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.5.4
PyYAML==5.1
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.5.2
segmentation-models-pytorch==0.3.0
six==1.16.0
stack-data==0.5.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
timm==0.4.12
tokenize-rt==5.0.0
tomli==2.0.1
torch==1.12.1
torchmetrics==0.10.2
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.2
tqdm==4.45.0
traitlets==5.5.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yacs==0.1.8
yarl==1.8.1
zipp==3.10.0
